Update optional notifications from email to Pushcut
Fixes #41.

Optional notifications so far have been sent with `mail`. While very convenient, this relies on configuration outside the workflow that often is out-of-control of the user. Also, it relies on the executable being available.

For these two reasons, Pushcut is now replacing `mail`. Pushcut requires minimal configuration outside the workflow that is within control of the user and only needs the `requests` Python lib. On the downside, Pushcut is proprietary and only available on iOS and iPadOS. I did not find an alternative that is free, easy to use, and better available.